signal_identification/requirements.txt
openai
python-dotenv
gradio
jupyter
ipykernel
